Fix image css by extracting the src only
The comic page has inlined css that causes the feed to look weird in
come clients as it uses min width and min height = 100%. In the previous
version, there was no inlined css.

This commit changes the regex to just extract the img src and then expose
feed items with an img block without any inlined css.